Move metadata_utils.py to filter_utils
