+++ matchms/filtering/metadata_processing/derive_smiles_from_inchi.py
@@ -1,7 +1,7 @@
 import logging
 from matchms.typing import SpectrumType
-from ...metadata_utils import (convert_inchi_to_smiles, is_valid_inchi,
-                               is_valid_smiles)
+from matchms.filtering.filter_utils.metadata_utils import (convert_inchi_to_smiles, is_valid_inchi,
+                                                           is_valid_smiles)
 
 
 logger = logging.getLogger("matchms")
