+++ tests/test_metadata_utils.py
@@ -5,18 +5,18 @@ from importlib.util import find_spec
 from unittest import mock
 import numpy as np
 import pytest
-import matchms.metadata_utils
-from matchms.metadata_utils import (derive_fingerprint_from_inchi,
-                                    derive_fingerprint_from_smiles,
-                                    is_valid_inchi, is_valid_inchikey,
-                                    is_valid_smiles, mol_converter)
+import matchms.filtering.filter_utils.metadata_utils
+from matchms.filtering.filter_utils.metadata_utils import (derive_fingerprint_from_inchi,
+                                                           derive_fingerprint_from_smiles,
+                                                           is_valid_inchi, is_valid_inchikey,
+                                                           is_valid_smiles, mol_converter)
 
 
 @pytest.fixture()
 def reload_metadata_utils():
     """Reload metadata_utils module after test has finished."""
     yield
-    reload(matchms.metadata_utils)
+    reload(matchms.filtering.filter_utils.metadata_utils)
 
 
 def test_mol_converter_smiles_to_inchi():
@@ -172,28 +172,28 @@ def test_missing_rdkit_module_error(reload_metadata_utils):
 
     expected_msg = "Conda package 'rdkit' is required for this functionality."
     with context:
-        reload(matchms.metadata_utils)
+        reload(matchms.filtering.filter_utils.metadata_utils)
         mol_input = "C[Si](Cn1cncn1)(c1ccc(F)cc1)"
         with pytest.raises(ImportError) as msg:
-            _ = matchms.metadata_utils.mol_converter(mol_input, "smiles", "inchikey")
+            _ = matchms.filtering.filter_utils.metadata_utils.mol_converter(mol_input, "smiles", "inchikey")
         assert expected_msg in str(msg.value), "Expected different ImportError."
 
         with pytest.raises(ImportError) as msg:
-            _ = matchms.metadata_utils.is_valid_inchi("test")
+            _ = matchms.filtering.filter_utils.metadata_utils.is_valid_inchi("test")
         assert expected_msg in str(msg.value), "Expected different ImportError."
 
         with pytest.raises(ImportError) as msg:
-            _ = matchms.metadata_utils.is_valid_smiles("test")
+            _ = matchms.filtering.filter_utils.metadata_utils.is_valid_smiles("test")
         assert expected_msg in str(msg.value), "Expected different ImportError."
 
         with pytest.raises(ImportError) as msg:
-            _ = matchms.metadata_utils.derive_fingerprint_from_inchi(mol_input, "test", 0)
+            _ = matchms.filtering.filter_utils.metadata_utils.derive_fingerprint_from_inchi(mol_input, "test", 0)
         assert expected_msg in str(msg.value), "Expected different ImportError."
 
         with pytest.raises(ImportError) as msg:
-            _ = matchms.metadata_utils.derive_fingerprint_from_smiles(mol_input, "test", 0)
+            _ = matchms.filtering.filter_utils.metadata_utils.derive_fingerprint_from_smiles(mol_input, "test", 0)
         assert expected_msg in str(msg.value), "Expected different ImportError."
 
         with pytest.raises(ImportError) as msg:
-            _ = matchms.metadata_utils.mol_to_fingerprint(mol_input, "test", 0)
+            _ = matchms.filtering.filter_utils.metadata_utils.mol_to_fingerprint(mol_input, "test", 0)
         assert expected_msg in str(msg.value), "Expected different ImportError."
